+++ vllm/transformers_utils/processors/__init__.py
@@ -11,6 +11,5 @@ reasons:
 from vllm.transformers_utils.processors.deepseek_vl2 import (
     DeepseekVLV2Processor)
 from vllm.transformers_utils.processors.ovis import OvisProcessor
-from vllm.transformers_utils.processors.step3_vl import Step3VisionProcessor
 
-__all__ = ["DeepseekVLV2Processor", "OvisProcessor", "Step3VisionProcessor"]
+__all__ = ["DeepseekVLV2Processor", "OvisProcessor"]
